+++ docs/en/docs/release-notes.md
@@ -7,11 +7,19 @@ hide:
 
 ## 0.20.2
 
+### Changed
+
+- Morph path argument into path option and expose it for all commands.
+
 ### Fixed
 
 - Properly detect wrapped Lilya instances.
 - Fix crash in show-urls.
 - Fix double initialization of app in runserver.
+
+### Breaking
+
+- lilya runserver loses its path argument. You can specify it via `lilya --path foo runserver`.
 
 ## 0.20.1